Minio support to substitute s3 on dev
This commit adds a minio.io container to docker-compose. This container
has the same API as s3 for writing/reading/etc, and it also runs on port
9000 of the Docker host.

Therefore this allows Insights app developers to use insights-upload
with this S3-compatible backend, instead of using S3 accounts (which
cost $$$), or localhost - which doesn't work for developing insights
apps as the consumers will need to fetch the file from a S3 bucket, not
from a file in a container.