Use payload_id established by 3Scale
3Scale will be assigning a unique ID to each upload, and the upload
service will now use that to track the payload through the platform.
This can be used by support to see where the payload is or where it
failed since all apps will identify it the same way.

The old 'hash' field is stil available for backward compatibility for
the time being.