change URL endpoints
This commit allows the API endpoint to be set via `APP_NAME` and
`PATH_PREFIX` environment variables. It also removes an extraneous
`/api/` from the middle of the URL.

Old URL to POST to: `/r/insights/platform/upload/api/v1/upload`

New URL to POST to: `/api/ingress/v1/upload`

If `APP_NAME` and `PATH_PREFIX` are not set, it will default to `/api/`
and `ingress`.

Additionally, the Pipfile has been updated to include `requests` and
`attrs`. These appeared to be missing.