Remove DEVMODE variable
commit_date and valid_topics should be more forgivable if either of
those things don't exist. We were doing this with DEVMODE but it's more
appropriate just to wrap them in try/except so that we can startup in a
dev environment even if we don't have these vars available.

This will particularly make local dev easier to get started.