Add leading zeros to self-generated request_id
If 3scale doesn't provide a request_id, or testing is done locally, we
generate a request_id. We also only reply with the proper header
containing that ID if it matches the leading zeroes format we have. The
chances of uuid4 generating a uuid that matches this check is super low