Fix the VALID_TOPICS built from /tmp/topics.json
The get_valid_topics() method builds a list of topic application_names
from a json file.  It was looping over the TOPIC_NAME splitting on the
'.' where the TOPIC_NAME was e.g. platform.upload.advisor instead of
taking the last array element.

This led to the VALID_TOPICS looking like:
```
['platform', 'upload', 'advisor', 'platform', 'upload',
 'advisor-pup', 'platform', 'upload', 'available', etc...]
```

Instead of:
```
['advisor', 'advisor-pup', 'available', etc...]
```